Splitting up MetricTaker
